Remove Applied AI ## Who I am:
- First name: Fredrick
- Last name: Onyango
- Email: ochieng.fredonyango@gmail.com
- LinkedIn profile: https://www.linkedin.com/in/fredonyango/

## I am interested about the position(s)

<!-- Check the postion(s) you are interested in -->
- [x] Technical mentor
- [x] Community Manager
- [x] Relevant Position

## My motivations

I am very motivated to want to join your team based on the engagement you're creating and defining the potential of Africa when in comes to technology. As a young tech trying to fit in this kind of ecosystem, this a place to build oneself and grow while identifying the hinderance as to why this potential of Africa going unnoticed.

## And I would like to add...

Have been in ALX since 2022, what struck me is things doesn't seem well as it appears. From the percentage of people who get jobs after graduation, to us who don't even get noticed. Let's find out why? This is the community.

## Final checklist

<!-- Make sure to check all these items -->

- [x] I add my resume (Markdown or PDF version) in the folder `applications`
- [x] I'm serious about my application
- [x] I'm ready to join the team and tacle incredible challenges
- [x] I'm matching the majority of the job requirements
- [x] I understand all positions are remote
- [x] All information provided about myself are true